requirements.txt
absl-py==0.6.1
astor==0.7.1
bleach==1.5.0
cap==0.0.114
cycler==0.10.0
Django==2.1.2
enum34==1.1.6
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
html5lib==0.9999999
image==1.5.27
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.1
nicer==0.0.36
numpy==1.14.5
opencv-python==3.4.3.18
Pillow==5.3.0
pkt==0.0.21
protobuf==3.6.1
pygame==1.9.4
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
tensorboard==1.10.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tflearn==0.3.2
threadsafe-tkinter==1.0.1
tqdm==4.28.1
typing==3.6.6
Werkzeug==0.14.1
